Adds good portion of file system support using IndexedDB/LocalStorage.
Updates webpack/babel to allow for async/await.

Still need to rewire some things and replace promises and some events
with async/await patterns.

Still need to wrap the Editor better with the intended Editor class.

Still need to add some documentation to the classes and generate the
documentation for the static site.