Updates label tracking to remove use of timer.
It parses the entire file when loaded and then parses changed labels
when the cursor moves onto and then off of a line. When the cursor moves
off of a line, it removes all labels on the previous version of the line
and adds labels that now exist on that line.