Adds patch to remove some other warnings from gdb.
These appear when the target is connected. So when gdb attaches to the
process being emulated. It warns about things being slow because it
thinks it is an actual serial connection transferring binary data. It's
all local network, so it is safe to ignore and is otherwise very
confusing to read.