Adds guards to prevent infinite loops and stepping after the simulator has quit.
GDB will sometimes get into a state where it no longer reads from input,
but can be invoked. This causes an infinite loop in the flush of stdin
since RAWRS was expecting a call into GDB to read at least one byte from
stdin. This stops the loop if it sees that nothing was consumed by the
call into GDB presuming that a subsequent call would not change matters.

This weird state must be fixed with GDB instead.