Allows `gdb` to control the breakpoints in the code listing.
This makes use of the `set breakpoint always-inserted on` command which
is invoked when `gdb` starts. This forces the breakpoints to be reported
to the target (our remote gdb server in Debugger) whenever they are
created. Otherwise, they are only reported lazily when the program is
about to be started via `s` and `c`, etc, commands.

When you set a breakpoint within `gdb`, it is reflected within the code
listing by checking the breakpoint box and highlighting the line. It
also removes the visual of the breakpoint when `gdb` deletes it.

The only other remaining task is to let `gdb` know when breakpoints
happen on the UI.