Allows UI breakpoint set/clear to be seen by `gdb`.
A quick and dirty but working way of getting breakpoints in and out of
gdb via the code listing UI. This invokes the `b` and `clear` commands
within `gdb`, which interrupts normal input, but it works for now. We
can expose the functions for invoking the commands we use directly in
order to not interrupt normal flow later.