Refactors file system and upload code.
The file upload is its own file that subclasses Dialog, which the float
explorer also subclasses.

LocalStorage is now FileSystem and manages the file storage via the
localForage library. This library handles the fallback to LocalStorage
for us and simplifies the API to basically what we had before. Great
minds n'at.

Restyles the file upload to be consistent with the rest of the program.

Fixes detection of a compiled binutils in `build.sh`.